Arreglos mensajes de error en delete categoria

+++ app/src/main/java/es/ucm/fdi/keeperly/repository/CategoriaRepository.java
@@ -11,11 +11,14 @@ import java.util.concurrent.Executors;
 
 import es.ucm.fdi.keeperly.data.local.database.KeeperlyDB;
 import es.ucm.fdi.keeperly.data.local.database.dao.CategoriaDAO;
+import es.ucm.fdi.keeperly.data.local.database.dao.PresupuestoDAO;
+import es.ucm.fdi.keeperly.data.local.database.dao.TransaccionDAO;
 import es.ucm.fdi.keeperly.data.local.database.entities.Categoria;
 
 public class CategoriaRepository {
 
     private final MutableLiveData<Integer> operationStatus = new MutableLiveData<>();
+    private final MutableLiveData<Integer> deleteStatus = new MutableLiveData<>();
 
 
     private final CategoriaDAO categoriaDao;
@@ -57,7 +60,25 @@ public class CategoriaRepository {
     }
 
     public void delete(Categoria categoria) {
-        executorService.execute(() -> categoriaDao.delete(categoria));
+        PresupuestoDAO presupuestoDAO = KeeperlyDB.getInstance().presupuestoDao();
+        TransaccionDAO transaccionDAO = KeeperlyDB.getInstance().transaccionDao();
+        deleteStatus.postValue(-1);
+        try {
+            if (!presupuestoDAO.getPresupuestosPorCategoria(categoria.getId()).isEmpty()) {
+                deleteStatus.postValue(-2);
+            } else {
+                if(transaccionDAO.getTransaccionesPorCategoria(categoria.getId()).isEmpty()) {
+                    executorService.execute(() -> categoriaDao.delete(categoria));
+                    deleteStatus.postValue(1);
+                }else{
+                    deleteStatus.postValue(-3);
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            deleteStatus.postValue(-1);
+        }
+
     }
 
     public LiveData<List<Categoria>> getAllCategorias() {
@@ -74,5 +95,9 @@ public class CategoriaRepository {
 
     public LiveData<Integer> getOperationStatus() {
         return operationStatus;
+    }
+
+    public LiveData<Integer> getDeleteStatus() {
+        return deleteStatus;
     }
 }
